Changed logout to use POST and minor bug fixes

+++ BankConfigurationPortal/BankConfigurationPortal.Web/Attributes/CookieAuthorizationAttribute.cs
@@ -51,6 +51,10 @@ namespace BankConfigurationPortal.Web.Attributes {
 
                 Session session = db.GetSession(cookieId);
 
+                if (session == null) {
+                    return false;
+                }
+
                 if (session.Expires <= DateTime.Now) {
                     db.DeleteSession(cookieId);
                     return false;
